Pull #705: separated classes based on logic versus api
